+++ rauthy-client/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## UNRELEASED
+
+### BREAKING
+
+The `rauthy_client::init()` has been renamed to `rauthy_client::init_with()` and a new `rauthy_client::init()`
+has been added which just uses safe defaults.
+
+### Changes
+
+- renamed `rauthy_client::init()` to `rauthy_client::init_with()` and added a new `rauthy_client::init()`
+  with safe defaults
+- added a new features `device-code` which will make it possible to use the Device Authorization Grant flow
+  with this client
+
 ## v0.3.0
 
 - Changes the `RauthyConfig` to now accept `ClaimMapping` for `admin_claim` and `user_claim`.